Java5 language level syntax migration

+++ src/test/java/io/reactivex/rxjava3/flowable/FlowableBackpressureTests.java
@@ -508,7 +508,7 @@ public class FlowableBackpressureTests extends RxJavaTest {
 
             System.out.println("testOnBackpressureDrop => Received: " + onNextEvents.size() + "  Emitted: " + c.get() + " Last value: " + lastEvent);
             // it drop, so we should get some number far higher than what would have sequentially incremented
-            assertTrue(num - 1 <= lastEvent.intValue());
+            assertTrue(num - 1 <= lastEvent);
         }
     }
 
@@ -548,7 +548,7 @@ public class FlowableBackpressureTests extends RxJavaTest {
             // in reality, num < passCount
             assertTrue(num <= passCount.get());
             // it drop, so we should get some number far higher than what would have sequentially incremented
-            assertTrue(num - 1 <= lastEvent.intValue());
+            assertTrue(num - 1 <= lastEvent);
             assertTrue(0 < dropCount.get());
             assertEquals(emitCount.get(), passCount.get() + dropCount.get());
         }
@@ -572,7 +572,7 @@ public class FlowableBackpressureTests extends RxJavaTest {
 
             System.out.println("testOnBackpressureDrop => Received: " + onNextEvents.size() + "  Emitted: " + c.get() + " Last value: " + lastEvent);
             // it drop, so we should get some number far higher than what would have sequentially incremented
-            assertTrue(num - 1 <= lastEvent.intValue());
+            assertTrue(num - 1 <= lastEvent);
         }
     }
 
@@ -600,7 +600,7 @@ public class FlowableBackpressureTests extends RxJavaTest {
 
             System.out.println("testOnBackpressureDrop => Received: " + onNextEvents.size() + " Dropped: " + dropCount.get() + "  Emitted: " + c.get() + " Last value: " + lastEvent);
             // it drop, so we should get some number far higher than what would have sequentially incremented
-            assertTrue(num - 1 <= lastEvent.intValue());
+            assertTrue(num - 1 <= lastEvent);
             // no drop in synchronous mode
             assertEquals(0, dropCount.get());
             assertEquals(c.get(), onNextEvents.size());
